+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `keepachangelog.release` function to guess new version number based on `Unreleased` section, update changelog and return new version number.
 - `keepachangelog.to_raw_dict` function returning a raw markdown description of the release under `raw` dict.
 
+### Fixed
+- Handle any category name.
+- Add more flexibility for release format.
+
+### Changed
+- `Unreleased` is now reported as lower cased `unreleased`.
+
 ## [0.4.0] - 2020-09-21
 ### Added
 - `keepachangelog.flask_restx.add_changelog_endpoint` function to add a changelog endpoint to a [Flask-RestX](https://flask-restx.readthedocs.io/en/latest/) application.